# perf ngClass batch add/remove classes operations