@atscott 2020-06-08
> @atscott, doesn't the existing tests already cover that multiple classes are added/removed to an element?
> 
> https://github.com/angular/angular/blob/650974e1b37fc1db6af93222a8eee25138e6492f/packages/common/test/directives/ng_class_spec.ts#L187-L195

Hmm, it sure does! I wonder if it would be useful to spy on the `Render2` `addClass`/`removeClass` and assert that it's only called once. 🤷