@JoostK 2020-06-08
Note also that this is a breaking change, so I'm afraid it can't go in atm.