@JoostK 2020-06-08
@alan-agius4  The `Renderer2`  interface is public API and it's expected to be implemented by third-parties, e.g. NativeScript. The changed method signature means that those implementors would need to be adjusted to deal with arrays, making this change breaking.