@alan-agius4 2020-06-10
Maybe we can use classList from the element directly in `ngClass` instead of using the renderer to avoid the breaking change?

Example from
```ts
this._renderer.removeClass(this._ngEl.nativeElement, klass);
```
to
```ts
this._ngEl.nativeElement.classList.remove(klass)
```