@pkozlowski-opensource 2020-06-10
> Maybe we can use classList from the element directly in ngClass instead of using the renderer to avoid the breaking change?

@alan-agius4 this would be a breaking change as well since the operations that were previously sent to a renderer all of the sudden wouldn't, no?