@alan-agius4 2020-06-10
@pkozlowski-opensource, indeed it can be if in the `addClass` you have some extra logic.