@alan-agius4 2020-06-18
@jelbourn, thanks for the feedback.

With regards to documentation about the improvements there isn't any. Mostly these improvements are for server side rendering were the DOM API is patched and add/remove operator is not optimised for repetitive operations. See: fgnass/domino#166. (While this helps this is not enough)

A 3P partner/client using Angular version 9 and Ivy reported that while they browser runtime became faster their server side runtime code is slower. Their tracked it down to `ngClass` with multiple classes.

Ex: `<div [ngClass]="{'class1 class2 class3’ : true}”></div>`

From there, myself and their engineers took 2 started to investigate the root cause, I did this on a toy app, while they did this on their real world app.

We both came to the same conclusion that the main bottleneck is that classes are added and removed individually, they did some patches to try to test what's the impact of added/removing classes all at once, and it resulted in 10% reduction of rendering time on the server.

Here are some matrix on my toy app for `<div [ngClass]="{'class1 class2 class3’ : true}”></div>` we get the following matrix on the server:

Current
```
_toggleClass 1.3ms
_applyKeyValueChanges 1.4ms
```

With this change
```
_toggleClass 0.9ms
_applyKeyValueChanges 1.0ms
```