@jelbourn 2020-06-19
@alan-agius4 Understood! Could you add a comment to the code that links to that domino issue as well as a very brief summary of that explanation about the client report? Everything else LGTM