@jelbourn 2020-07-02
Based on @pkozlowski-opensource's comments, I'm also wondering whether it's worth doing this now. How hard would it be to fix on the domino side?