@alan-agius4 2020-07-02
@pkozlowski-opensource & @jelbourn, thanks for the feedback. 

There is already a PR for domino (https://github.com/fgnass/domino/pull/166) to improve the performance related to repeated operations. Despite of that, since the framework is invoking `addClass`, `removeClass` on the same node multiple times (One for each class on a given Node) there is still some overhead over calling addClass/removeClass with an array.  I will actually gather some more timings and numbers around the difference between this change and Dominos change. Likely though this will have to wait some time, as currently I am mostly focused on v10 issues, and seeing that this is for v11, there is no immediate urgency in tackling this right away.

That being said, in general I do think that irrespective of this fix/change, I think it is good to allow the renderer to add/remove multiple classes at once since this is part of the standard DOM API.

With regards to @pkozlowski-opensource comment about the deprecation and removal of NgClass/NgStyle, I am aware of that, but to completely remove those will take several major versions, and I am hoping that the alternate approach to these 2 directives will take into more into consideration also performance with mocked DOM APIs.

Side note: recently we also had chats around the possibility in replacing Domino with something which is more maintained and supported. However, from some initial benchmarking it still looks like Domino is one of the faster Node.JS DOM APIs.

@mgechev, can you kindly try to reach out to Domino and get https://github.com/fgnass/domino/pull/166 merged?