@alan-agius4 2020-09-18
Closing as the performance fix from Domino has landed, which makes the timings acceptable.