@alfaproject 2020-09-18
Wouldn't this still be a decent addition to the API?